feat(hpu): Implement new version of the Hpu Instruction Set
In depth rework of the DOp/IOp definition.
Dop structure was almost the same, however, the asm definition/parser
were rework with the following objectives:
* Template as first class citizen
* Support of Immediate template
* Direct parser and conversion between Asm/Hex
* Replace deku (and it's associated endianess limitation) by
  bitfield_struct and manual parsing

IOp structure was deeply rework. The aims was to introduce more
flexibiliy.
* Support various number of Destination
* Support various number of Sources
* Support various number of Immediat values
* Support of multiple bitwidth (Not implemented yet in the Fpga
  firmware)

Add lot of information in the IOp prototype that could be used at
runtime for proper error handling.

+++ backends/tfhe-hpu-backend/src/interface/rtl/runtime.rs
@@ -118,7 +118,7 @@ impl InfoPePbs {
             .register()
             .get("Bpip::timeout")
             .expect("Unknow register, check regmap definition");
-        self.bpip_timeout = ffi_hw.read_reg(*reg_timeout.offset() as u64) as u32;
+        self.bpip_timeout = ffi_hw.read_reg(*reg_timeout.offset() as u64);
     }
 
     pub fn update_loop(&mut self, ffi_hw: &mut ffi::HpuHw, regmap: &FlatRegmap) {
